Fix handling of multipart/mixed emails
Previously, ingest-file would add the body of HTML parts to the `bodyHtml property and the body of plaintext parts to the `bodyText` property.

Given an email with the following structure…

```
multipart/mixed
    text/plain
    text/html
    text/plain
    text/html
```

… this would result in an FtM entity that has two values in the `bodyHtml` and two values in the `bodyText` property with no way of reconstructing the order of all four parts in the original email.

In case of `multipart/mixed` emails, we need to display all parts in their original order (no matter the content type of the individual parts) to provide a complete, correct preview of the email.

I have solved this by generating alternative HTML or plaintext representation for the parts of `multipart/mixed` emails and adding them to the respective entity properties.

* If a mixed email has a plaintext part, the text is HTML encoded and line breaks are converted to `<br>` tags.
* If a mixed email has an HTML part, all HTML markup is stripped (we already do this for search).

This approach adds some storage overhead as we need to store the contents of email twice. But that is the whole idea of multipart emails with alternative representations… Also, the text contents of most emails are rather short.

Fixes #3161