Refactor to use explicit recursion
Hiding the recursion in the `walk_parts` iterator is a bit pointless when we also do that explicitly in `parse_part`. Makes the code much easier to follow in my opinion and allows us to stop the traversal earlier, e.g. when we encounter an attachment. The latter is key for fixing #3164.