Do not inline attached emails
While ingest-file already handles parts of multipart emails with the `Content-Disposition: attachment` header, the current implementation fails for attached emails. This is due to the common structure of emails with attached emails:

Structure of an email with an attached plaintext email:

```
multipart/mixed
    text/plain
    message/rfc822
        text/plain
```

Structure of an email with an attached multipart email:

```
multipart/mixed
    text/plain
    message/rfc822
        multipart/alternative
            text/plain
            text/html
```

In both cases, the `Content-Disposition` header will be set only on the `message/rfc822` part. The previous implementation would simply skip this part because it also is a multipart part itself. The iterator would then descend into its subparts and handle the `text/plain` part (in the first example) or the `mutlipart/alternative` part (in the second example). As the subparts do not have the `Content-Disposition` header set, they would be handled like any other part of the original email.

The new implementation first checks whether a part is an attachment. If it is, it is processed as an attachment and we do not continue the recursion (to prevent the "inlining" of subparts of the attachment).

Fix #3164.