Display headers of attached inline emails
Emails can also contains attached inline emails. This is common especially for automated delivery notifications, but some email clients seem to use attached inline emails when replying to another email. Most common email client display relevant headers of the attached inline emails. This change appends a text-representation of relevant email headers to the `bodyText` and `bodyHtml` properties.

While a better solution might be to store the parsed email contents in a more structured way, we're a bit limit by FtM and this seems like a pragmatic solution for a case that is probably rather uncommon.